# specs/001-feature-name-cv/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CV Optimization API
  description: API for CV upload, parsing, editing, and AI-enhanced optimization
  version: 1.0.0
  contact:
    name: CV Optimization Team
    email: support@cvoptimizer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.cvoptimizer.com
    description: Production server
security:
  - BearerAuth: []
  - OAuth2: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshToken'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cvs/{cv_id}/generate-section:
    post:
      tags:
        - AI Features
      summary: Generate AI section
      description: Generate AI-enhanced section for CV based on job description
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIGenerationRequest'
      responses:
        '200':
          description: AI section generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISectionResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: CV or job description not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cvs/{cv_id}/ai-sections:
    get:
      tags:
        - AI Features
      summary: Get AI sections
      description: Get all AI-generated sections for a CV
      parameters:
        - name: cv_id
          in: path
          required: true
          description: CV identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI sections retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISectionListResponse'
        '404':
          description: CV not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: Email access
            profile: Profile access
  schemas:
...
    UserRegistration:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password
        full_name:
          type: string
          description: User's full name
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    RefreshToken:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'

    # User Schema
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time

    # CV Schemas
...
    AIGenerationRequest:
      type: object
      required:
        - job_description_id
      properties:
        job_description_id:
          type: string
          format: uuid
          description: Job description to use for generation
        section_type:
          type: string
          enum: [why_good_fit, cover_letter, skills_highlight]
          default: why_good_fit
          description: Type of section to generate
    AISectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cv_id:
          type: string
          format: uuid
        job_description_id:
          type: string
          format: uuid
        section_content:
          type: string
        section_type:
          type: string
        model_used:
          type: string
        tokens_used:
          type: integer
        generation_time:
          type: integer
        created_at:
          type: string
          format: date-time
    AISectionListResponse:
      type: object
      properties:
        ai_sections:
          type: array
          items:
            $ref: '#/components/schemas/AISectionResponse'

    # Common Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        success:
          type: boolean
          default: false